opaecxx: consolidate exception classes
* Move exception classes that correspond to fpga_result codes to except.h
* Add unit tests for using the macro for each exception class
