+++ common/include/opae/cxx/core/except.h
@@ -29,6 +29,7 @@
 
 #include <opae/types_enum.h>
 
+
 namespace opae {
 namespace fpga {
 namespace types {
@@ -63,11 +64,12 @@ class src_location {
   int line_;
 };
 
+
 /// Construct a src_location object for the current source line.
 #define OPAECXX_HERE \
   opae::fpga::types::src_location(__FILE__, __func__, __LINE__)
 
-/** Runtime exception
+/** Generic OPAE exception
  *
  * An except tracks the source line of origin
  * and an optional fpga_result. If no fpga_result
@@ -87,7 +89,6 @@ class except : public std::exception {
   /** except constructor
    *
    * @param[in] res The fpga_result value associated with this exception.
-   * @param[in] msg The error message as a string
    * @param[in] loc Location where the exception was constructed.
    */
   except(fpga_result res, src_location loc) noexcept;
@@ -115,6 +116,254 @@ class except : public std::exception {
   mutable char buf_[MAX_EXCEPT];
 };
 
+/** invalid_param exception
+ *
+ * invalid_param tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_INVALID_PARAM is returned from a call to
+ * an OPAE C API function
+ */
+class invalid_param : public except {
+  public:
+
+    /** invalid_param constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    invalid_param(src_location loc) noexcept
+    : except(FPGA_INVALID_PARAM, "failed with return code FPGA_INVALID_PARAM", loc)
+    {}
+
+};
+
+/** busy exception
+ *
+ * busy tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_BUSY is returned from a call to
+ * an OPAE C API function
+ */
+class busy : public except {
+  public:
+
+    /** busy constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    busy(src_location loc) noexcept
+    : except(FPGA_BUSY, "failed with return code FPGA_BUSY", loc)
+    {}
+
+};
+
+/** exception exception
+ *
+ * exception tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_EXCEPTION is returned from a call to
+ * an OPAE C API function
+ */
+class exception : public except {
+  public:
+
+    /** exception constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    exception(src_location loc) noexcept
+    : except(FPGA_EXCEPTION, "failed with return code FPGA_EXCEPTION", loc)
+    {}
+
+};
+
+/** not_found exception
+ *
+ * not_found tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_NOT_FOUND is returned from a call to
+ * an OPAE C API function
+ */
+class not_found : public except {
+  public:
+
+    /** not_found constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    not_found(src_location loc) noexcept
+    : except(FPGA_NOT_FOUND, "failed with return code FPGA_NOT_FOUND", loc)
+    {}
+
+};
+
+/** no_memory exception
+ *
+ * no_memory tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_NO_MEMORY is returned from a call to
+ * an OPAE C API function
+ */
+class no_memory : public except {
+  public:
+
+    /** no_memory constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    no_memory(src_location loc) noexcept
+    : except(FPGA_NO_MEMORY, "failed with return code FPGA_NO_MEMORY", loc)
+    {}
+
+};
+
+/** not_supported exception
+ *
+ * not_supported tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_NOT_SUPPORTED is returned from a call to
+ * an OPAE C API function
+ */
+class not_supported : public except {
+  public:
+
+    /** not_supported constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    not_supported(src_location loc) noexcept
+    : except(FPGA_NOT_SUPPORTED, "failed with return code FPGA_NOT_SUPPORTED", loc)
+    {}
+
+};
+
+/** no_driver exception
+ *
+ * no_driver tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_NO_DRIVER is returned from a call to
+ * an OPAE C API function
+ */
+class no_driver : public except {
+  public:
+
+    /** no_driver constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    no_driver(src_location loc) noexcept
+    : except(FPGA_NO_DRIVER, "failed with return code FPGA_NO_DRIVER", loc)
+    {}
+
+};
+
+/** no_daemon exception
+ *
+ * no_daemon tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_NO_DAEMON is returned from a call to
+ * an OPAE C API function
+ */
+class no_daemon : public except {
+  public:
+
+    /** no_daemon constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    no_daemon(src_location loc) noexcept
+    : except(FPGA_NO_DAEMON, "failed with return code FPGA_NO_DAEMON", loc)
+    {}
+
+};
+
+/** no_access exception
+ *
+ * no_access tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_NO_ACCESS is returned from a call to
+ * an OPAE C API function
+ */
+class no_access : public except {
+  public:
+
+    /** no_access constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    no_access(src_location loc) noexcept
+    : except(FPGA_NO_ACCESS, "failed with return code FPGA_NO_ACCESS", loc)
+    {}
+
+};
+
+/** reconf_error exception
+ *
+ * reconf_error tracks the source line of origin
+ * for exceptions thrown when the error code
+ * FPGA_RECONF_ERROR is returned from a call to
+ * an OPAE C API function
+ */
+class reconf_error : public except {
+  public:
+
+    /** reconf_error constructor
+     *
+     * @param[in] loc Location where the exception was constructed.
+     */
+    reconf_error(src_location loc) noexcept
+    : except(FPGA_RECONF_ERROR, "failed with return code FPGA_RECONF_ERROR", loc)
+    {}
+
+};
+
+namespace detail {
+
+/** typedef function pointer that returns bool if result is FPGA_OK
+ */
+typedef bool (*exception_fn)(fpga_result, const opae::fpga::types::src_location & loc);
+
+/** is_ok is a template function that throws an excpetion of its template argument type
+ *  if the result code is not FPGA_OK. Otherwise it returns true.
+ */
+template<typename T>
+constexpr bool is_ok(fpga_result result, const opae::fpga::types::src_location & loc) {
+  return result == FPGA_OK ? true : throw T(loc);
+}
+
+
+static exception_fn opae_exceptions[12] = { is_ok<opae::fpga::types::invalid_param>,
+                                            is_ok<opae::fpga::types::busy>,
+                                            is_ok<opae::fpga::types::exception>,
+                                            is_ok<opae::fpga::types::not_found>,
+                                            is_ok<opae::fpga::types::no_memory>,
+                                            is_ok<opae::fpga::types::not_supported>,
+                                            is_ok<opae::fpga::types::no_driver>,
+                                            is_ok<opae::fpga::types::no_daemon>,
+                                            is_ok<opae::fpga::types::no_access>,
+                                            is_ok<opae::fpga::types::reconf_error> };
+
+
+static inline void assert_fpga_ok(fpga_result result,
+                                  const opae::fpga::types::src_location & loc) {
+  if (result >= FPGA_OK && result <= FPGA_RECONF_ERROR)
+    // our exception table above starts at invalid_param with index 0
+    // but FPGA_INVALID_PARAM is actually enum 1 - let's account for that
+    opae_exceptions[result-1](result, loc);
+}
+
+}  // end of namespace detail
+
+/// Macro to check of result is FPGA_OK
+/// If not, throw exception that corresponds
+/// to the result code
+#define ASSERT_FPGA_OK(r) \
+  opae::fpga::types::detail::assert_fpga_ok(r, \
+      opae::fpga::types::src_location(__FILE__, __func__, __LINE__));
+
 }  // end of namespace types
 }  // end of namespace fpga
 }  // end of namespace opae
+
+
+
+
