--- common/include/opae/cxx/errors/not_supported.h
@@ -1,58 +1,0 @@
-// Copyright(c) 2018, Intel Corporation
-//
-// Redistribution  and  use  in source  and  binary  forms,  with  or  without
-// modification, are permitted provided that the following conditions are met:
-//
-// * Redistributions of  source code  must retain the  above copyright notice,
-//   this list of conditions and the following disclaimer.
-// * Redistributions in binary form must reproduce the above copyright notice,
-//   this list of conditions and the following disclaimer in the documentation
-//   and/or other materials provided with the distribution.
-// * Neither the name  of Intel Corporation  nor the names of its contributors
-//   may be used to  endorse or promote  products derived  from this  software
-//   without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING,  BUT NOT LIMITED TO,  THE
-// IMPLIED WARRANTIES OF  MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED.  IN NO EVENT  SHALL THE COPYRIGHT OWNER  OR CONTRIBUTORS BE
-// LIABLE  FOR  ANY  DIRECT,  INDIRECT,  INCIDENTAL,  SPECIAL,  EXEMPLARY,  OR
-// CONSEQUENTIAL  DAMAGES  (INCLUDING,  BUT  NOT LIMITED  TO,  PROCUREMENT  OF
-// SUBSTITUTE GOODS OR SERVICES;  LOSS OF USE,  DATA, OR PROFITS;  OR BUSINESS
-// INTERRUPTION)  HOWEVER CAUSED  AND ON ANY THEORY  OF LIABILITY,  WHETHER IN
-// CONTRACT,  STRICT LIABILITY,  OR TORT  (INCLUDING NEGLIGENCE  OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,  EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
-#pragma once
-#include <cstdint>
-#include <exception>
-
-#include <opae/types_enum.h>
-
-namespace opae {
-namespace fpga {
-namespace types {
-
-/** not_supported exception
- *
- * not_supported tracks the source line of origin
- * for exceptions thrown when the error code
- * FPGA_NOT_SUPPORTED is returned from a call to
- * an OPAE C API function
- */
-class not_supported : public except {
-  public:
-
-    /** not_supported constructor
-     *
-     * @param[in] loc Location where the exception was constructed.
-     */
-    not_supported(src_location loc) noexcept
-    : except(FPGA_NOT_SUPPORTED, "failed with return code FPGA_NOT_SUPPORTED", loc)
-    {}
-
-};
-
-}  // end of namespace types
-}  // end of namespace fpga
-}  // end of namespace opae
